Fix bug if input is not extacly power of 2
If the input value is not exatcly power of 2, the function will
return incorrect result. Firstly, no need to do  x-- in the begining
if x is power of 2. We treate every value as the same. Therefore, we don't
need to tailed "+1" at the end of return since 'x--' was removed in the
begining for special condition.

When the calculation range comes to x < 0xf, firstly, the function check
if x > 0x3. After that, the remaining part of x become 3 conditions,
which are 01'b, 10'b and 11'b. For 10'b and 11'b, we have to add 1 for the
result and for 01'b, the result stay the same.