Enhance the hamming distance algorithm
By observing with bit-level perspective, we could make below conclusion
For all the nums, we calculate how many 1's at each num's bit level.
For example,
if there is not 1 in bit 0 for all nums, hamming distance = 0.
if there is a 1'b in bit 0 for all nums, Hamming Distance (1)* (numsize - 1)
if there is two 1'b in bit 0 for all nums, Hamming Distance (2)* (numsize - 2)

By using the above sequence, we create a bit_array to calculate and store all the
numbers of 1'b in bit0=bit32 for all nums and calculate the total hamming distance
via above method.